refactor DailyForm
